add new category, Logic, and separate from Value
